+++ vulkanalia-sys/src/constants.rs
@@ -27,6 +27,7 @@ pub const MAX_GLOBAL_PRIORITY_SIZE_KHR: usize = 16;
 pub const MAX_MEMORY_HEAPS: usize = 16;
 pub const MAX_MEMORY_TYPES: usize = 32;
 pub const MAX_PHYSICAL_DEVICE_NAME_SIZE: usize = 256;
+pub const MAX_SHADER_MODULE_IDENTIFIER_SIZE_EXT: usize = 32;
 pub const QUEUE_FAMILY_EXTERNAL: u32 = !1;
 pub const QUEUE_FAMILY_FOREIGN_EXT: u32 = !2;
 pub const QUEUE_FAMILY_IGNORED: u32 = !0;